refactor: Remove references to ByteCode in the compiler
