+++ compiler/include/compiler/backend/code_writer.hpp
@@ -3,7 +3,8 @@
 
 #include <ostream>
 
-#include "bytecode/byte_code.hpp"
+#include "bytecode/code_chunk.hpp"
+#include "bytecode/header.hpp"
 #include "compiler/backend/constant.hpp"
 
 namespace fluir {
@@ -12,10 +13,7 @@ namespace fluir {
     explicit CodeWriter(std::ostream&);
     virtual ~CodeWriter() = default;
 
-    void write(const code::ByteCode& code, std::ostream& destination);
-
     virtual void writeHeader(const code::Header&) = 0;
-    virtual void writeConstants(const std::vector<code::Value>&) = 0;
     virtual void writeConstants(const be::ConstantsArray&) = 0;
     virtual void writeChunk(const code::Chunk&) = 0;
